+++ src/background/pipeline/__tests__/RecordingPipeline-r2.test.ts
@@ -41,7 +41,7 @@ vi.mock('../../../utils/errorUtils.js', () => ({
 vi.mock('../../../utils/pendingStorage.js', () => ({
   addPendingPage: vi.fn().mockResolvedValue(undefined),
 }));
-vi.mock('../../../utils/crypto.js', () => ({ hashUrl: vi.fn().mockResolvedValue('mocked-hash') }));
+vi.mock('../../../utils/crypto/index.js', () => ({ hashUrl: vi.fn().mockResolvedValue('mocked-hash') }));
 vi.mock('../../../utils/storageUrls.js');
 vi.mock('../../../utils/domainUtils.js');
 vi.mock('../../../utils/permissionManager.js');
