+++ src/utils/settingsExportImport.ts
@@ -4,8 +4,8 @@
  */
 
 import { getSettings, saveSettings, getOrCreateHmacSecret, Settings, API_KEY_FIELDS } from './storage.js';
-import { computeHMAC, encrypt, decryptData, deriveKey } from './crypto.js';
-import { generateSalt } from './crypto.js';
+import { computeHMAC, encrypt, decryptData, deriveKey } from './crypto/index.js';
+import { generateSalt } from './crypto/index.js';
 import { logError, logWarn, logInfo, ErrorCode } from './logger.js';
 import { errorMessage } from './errorUtils.js';
