+++ src/utils/storage/settingsStore.ts
@@ -7,7 +7,7 @@
 import { logInfo, logDebug, logError, ErrorCode } from '../logger.js';
 import { errorMessage } from '../errorUtils.js';
 import { migrateUblockSettings, migrateJpLayoutDefault, migrateCategoryBDefault, migrateWhitelistExtractionDefault } from '../migration.js';
-import { isEncrypted, encryptApiKey, decryptApiKey } from '../crypto.js';
+import { isEncrypted, encryptApiKey, decryptApiKey } from '../crypto/index.js';
 import { withOptimisticLock } from '../optimisticLock.js';
 import { normalizeUrl } from '../urlUtils.js';
 import { getOrCreateEncryptionKey } from './encryptionSession.js';
